200-other/201-strategic-audit-bot/requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
tavily-python>=0.3.0
jinja2>=3.1.0
markdown>=3.4.0
fpdf2>=2.7.0
python-dotenv>=1.0.0

